Close current item @click on link in vertical menu
Prior to this commit an opened item would only get closed if the route-to and the current route are different, e.g. when the user goes to "/inbox" and clicks on a question, then a second click on "Unbeantwortet" would have no effect because the routes are identical.

With this commit a click on any label or state will always close the current item and return to the list view.