## Background The Back and Forward buttons in the questionnaire always refer to a specific context (unanswered, answered, marked). 
If a user answers an unanswered question and moves on to the next one, the question just answered cannot be called up with the back button, because it is now in a different context.

To spare users unnecessary wonder why the back button does not lead to the previous question, the back button is greyed out if no going back is possible.

## Technical
### `Email.vue`
- `isFirstMailResult` is initiated in `data()`, and …
- updated in `nextMail()`, `previousMail()` and `updateOpenMail()`, and …
- passed as a prop `isFirstMail` to `EmailView.vue`

### `EmailView.vue`
- `EmailView.vue` receives the prop `isFirstMail` from `Email.vue`, and …
- sets class `text-grey` if `isFirstMail` is `true`

Fixes IELSI-36